slimbus: qcom-ngd-ctrl: Check the state of LPASS when probing
When relying on getting notified of SSR events from LPASS to start, slimbus
never starts if LPASS is already powered on when qcom-ngd-ctrl is probed.
Check the current state of LPASS when probing and start directly if it is
already on.

Also move initializing ngd_up_work before anything could possibly schedule it
while uninitialized.

Signed-off-by: Yassine Oudjana <y.oudjana@protonmail.com>